:truck: move logging to utils path
:truck: make sure script points to new logging location
